basic dungeon functionality

+++ utils/datalocks.js
@@ -8,6 +8,7 @@ module.exports = {
         return fs.existsSync(path.join(__dirname, `../userdata/.lock ${userId}`));
     },
     unlockData: function(userId) {
-        fs.rmSync(path.join(__dirname, `../userdata/.lock ${userId}`));
+        if (fs.existsSync(path.join(__dirname, `../userdata/.lock ${userId}`)))
+            fs.rmSync(path.join(__dirname, `../userdata/.lock ${userId}`));
     }
 }
